net: phy: inphi in112525 drop flash vs eMMC
Instead of using two hard coded cases from compilation, let's have a
runtime setting.
It allows to share the same uboot code whenever it runs from a nand/nor
flash or from the eMMC/sd card.

I did not try to fix up the ident in order to minimize the diff of code.

If it looks OK, it should be rebased.

Signed-off-by: Vincent Jardin <vjardin@free.fr>